[Rewards] Move onboaring from overlays to panel.
