+++ components/brave_rewards/resources/rewards_panel/components/panel.style.ts
@@ -1,0 +1,10 @@
+/* Copyright (c) 2023 The Brave Authors. All rights reserved.
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this file,
+ * You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+import styled from 'styled-components'
+
+export const panel = styled.div`
+  padding: 13px;
+`
