+++ components/brave_rewards/resources/rewards_panel/components/panel.tsx
@@ -6,17 +6,24 @@
 import * as React from 'react'
 
 import { HostContext, useHostListener } from '../lib/host_context'
-import { getProviderPayoutStatus } from '../../shared/lib/provider_payout_status'
+import { TabOpenerContext } from '../../shared/components/new_tab_link'
+import { OnboardingResult, RewardsOptIn } from '../../shared/components/onboarding'
 import { WalletCard } from '../../shared/components/wallet_card'
+import { getProviderPayoutStatus } from '../../shared/lib/provider_payout_status'
 import { LimitedView } from './limited_view'
 import { NavBar } from './navbar'
 import { PanelOverlays } from './panel_overlays'
 import { PublisherCard } from './publisher_card'
 
+import * as urls from '../../shared/lib/rewards_urls'
+
+import * as style from './panel.style'
+
 type ActiveView = 'tip' | 'summary'
 
 export function Panel () {
   const host = React.useContext(HostContext)
+  const tabOpener = React.useContext(TabOpenerContext)
 
   const [userType, setUserType] = React.useState(host.state.userType)
   const [balance, setBalance] = React.useState(host.state.balance)
@@ -36,6 +43,19 @@ export function Panel () {
   const [activeView, setActiveView] = React.useState<ActiveView>(
     publisherInfo ? 'tip' : 'summary')
 
+  const [requestedView, setRequestedView] =
+    React.useState(host.state.requestedView)
+  const [availableCountries, setAvailableCountries] =
+    React.useState(host.state.availableCountries)
+  const [defaultCountry, setDefaultCountry] =
+    React.useState(host.state.defaultCountry)
+  const [onboardingResult, setOnboardingResult] =
+    React.useState<OnboardingResult | null>(null)
+  const [rewardsEnabled, setRewardsEnabled] =
+    React.useState(host.state.rewardsEnabled)
+  const [declaredCountry, setDeclaredCountry] =
+    React.useState(host.state.declaredCountry)
+
   useHostListener(host, (state) => {
     setUserType(state.userType)
     setBalance(state.balance)
@@ -46,7 +66,45 @@ export function Panel () {
     setPayoutStatus(state.payoutStatus)
     setSummaryData(state.summaryData)
     setPublisherInfo(state.publisherInfo)
+
+    setRequestedView(state.requestedView)
+    setRewardsEnabled(state.rewardsEnabled)
+    setDeclaredCountry(state.declaredCountry)
+    setAvailableCountries(state.availableCountries)
+    setDefaultCountry(state.defaultCountry)
   })
+
+  const needsCountry = rewardsEnabled && !declaredCountry
+
+  function renderOnboaring () {
+    const onHideResult = () => {
+      if (onboardingResult === 'success') {
+        setTimeout(() => {
+          tabOpener.openTab(urls.rewardsTourURL)
+        }, 1000)
+      }
+      setOnboardingResult(null)
+      host.closePanel()
+    }
+
+    const onEnable = (country: string) => {
+      host.enableRewards(country).then((result) => {
+        setOnboardingResult(result)
+      })
+    }
+
+    return (
+      <RewardsOptIn
+        availableCountries={availableCountries}
+        defaultCountry={defaultCountry}
+        initialView={needsCountry || requestedView === 'rewards-setup' ?
+          'declare-country' : 'default'}
+        result={onboardingResult}
+        onEnable={onEnable}
+        onHideResult={onHideResult}
+      />
+    )
+  }
 
   const walletProvider = externalWallet ? externalWallet.provider : null
 
@@ -84,11 +142,15 @@ export function Panel () {
   }
 
   return (
-    <div>
-      <div className='rewards-panel' data-test-id='rewards-panel'>
-        {userType !== 'unconnected' ? renderFull() : <LimitedView />}
-      </div>
-      <PanelOverlays />
-    </div>
+    onboardingResult || !rewardsEnabled || needsCountry
+      ? <div className='rewards-panel' data-test-id='rewards-panel'>
+          {renderOnboaring()}
+        </div>
+      : <style.panel>
+          <div className='rewards-panel' data-test-id='rewards-panel'>
+            {userType !== 'unconnected' ? renderFull() : <LimitedView />}
+          </div>
+          <PanelOverlays />
+        </style.panel>
   )
 }
