+++ components/brave_rewards/resources/rewards_panel/components/panel_overlays.tsx
@@ -6,30 +6,15 @@
 import * as React from 'react'
 
 import { HostContext, useHostListener } from '../lib/host_context'
-import { OnboardingResult, RewardsOptInModal } from '../../shared/components/onboarding'
 import { AdaptiveCaptchaView } from '../../rewards_panel/components/adaptive_captcha_view'
 import { GrantCaptchaModal } from './grant_captcha_modal'
 import { NotificationOverlay } from './notification_overlay'
 import { VBATNoticeModal } from './vbat_notice_modal'
 import { shouldShowVBATNotice } from '../../shared/components/vbat_notice'
-import { TabOpenerContext } from '../../shared/components/new_tab_link'
-
-import * as urls from '../../shared/lib/rewards_urls'
 
 export function PanelOverlays() {
   const host = React.useContext(HostContext)
-  const tabOpener = React.useContext(TabOpenerContext)
 
-  const [requestedView, setRequestedView] =
-    React.useState(host.state.requestedView)
-  const [rewardsEnabled, setRewardsEnabled] =
-    React.useState(host.state.rewardsEnabled)
-  const [declaredCountry, setDeclaredCountry] =
-    React.useState(host.state.declaredCountry)
-  const [availableCountries, setAvailableCountries] =
-    React.useState(host.state.availableCountries)
-  const [defaultCountry, setDefaultCountry] =
-    React.useState(host.state.defaultCountry)
   const [options, setOptions] = React.useState(host.state.options)
   const [grantCaptchaInfo, setGrantCaptchaInfo] =
     React.useState(host.state.grantCaptchaInfo)
@@ -40,53 +25,14 @@ export function PanelOverlays() {
   const [userType, setUserType] = React.useState(host.state.userType)
   const [notificationsHidden, setNotificationsHidden] = React.useState(false)
   const [hideVBATNotice, setHideVBATNotice] = React.useState(false)
-  const [onboardingResult, setOnboardingResult] =
-    React.useState<OnboardingResult | null>(null)
-
-  const needsCountry = rewardsEnabled && !declaredCountry
 
   useHostListener(host, (state) => {
-    setRequestedView(state.requestedView)
-    setRewardsEnabled(state.rewardsEnabled)
-    setDeclaredCountry(state.declaredCountry)
-    setAvailableCountries(state.availableCountries)
-    setDefaultCountry(state.defaultCountry)
     setOptions(state.options)
     setGrantCaptchaInfo(state.grantCaptchaInfo)
     setNotifications(state.notifications)
     setAdaptiveCaptchaInfo(state.adaptiveCaptchaInfo)
     setUserType(state.userType)
   })
-
-  if (onboardingResult || !rewardsEnabled || needsCountry) {
-    const onHideResult = () => {
-      if (onboardingResult === 'success') {
-        setTimeout(() => {
-          tabOpener.openTab(urls.rewardsTourURL)
-        }, 1000)
-      }
-      setOnboardingResult(null)
-      host.closePanel()
-    }
-
-    const onEnable = (country: string) => {
-      host.enableRewards(country).then((result) => {
-        setOnboardingResult(result)
-      })
-    }
-
-    return (
-      <RewardsOptInModal
-        availableCountries={availableCountries}
-        defaultCountry={defaultCountry}
-        initialView={needsCountry || requestedView === 'rewards-setup' ?
-          'declare-country' : 'default'}
-        result={onboardingResult}
-        onEnable={onEnable}
-        onHideResult={onHideResult}
-      />
-    )
-  }
 
   if (adaptiveCaptchaInfo) {
     const onContactSupport = () => {
