+++ components/brave_rewards/resources/shared/components/onboarding/index.ts
@@ -2,5 +2,5 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-export { RewardsOptInModal, OnboardingResult } from './rewards_opt_in_modal'
+export { RewardsOptIn, OnboardingResult } from './rewards_opt_in'
 export { SettingsOptInForm } from './settings_opt_in_form'
